requirements.txt
## The following requirements were added by pip freeze:
amqp==2.5.1
appdirs==1.4.3
attrs==19.1.0
Automat==0.7.0
billiard==3.6.1.0
btrfsutil==1.1.1
CacheControl==0.12.5
celery==4.3.0
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
defusedxml==0.6.0
distlib==0.2.9
distro==1.4.0
Django==2.2.5
django-allauth==0.40.0
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-rest-auth==0.9.5
django-timezone-field==3.0
djangorestframework==3.10.3
dnspython==1.16.0
eventlet==0.25.1
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
kombu==4.6.4
libmsym==0.2.4
lit==0.8.1.dev0
lockfile==0.12.2
monotonic==1.5
more-itertools==4.3.0
msgpack==0.6.1
oauthlib==3.1.0
optimus-manager==1.2
packaging==19.1
pep517==0.6.0
progress==1.5
pwquality==1.4.0
PyHamcrest==1.9.0
pyparsing==2.4.2
PyQt5==5.13.1
PyQt5-sip==4.19.19
python-crontab==2.3.9
python-dateutil==2.8.0
python3-openid==3.1.0
pytoml==0.1.21
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
retrying==1.3.3
six==1.12.0
sqlparse==0.3.0
team==1.0
Twisted==19.7.0
urllib3==1.25.3
vine==1.3.0
webencodings==0.5.1
zipp==0.6.0
zope.interface==4.6.0
